+++ src/app/api/openai/route.ts
@@ -2,8 +2,7 @@ import { headers } from "next/headers";
 import { NextResponse } from "next/server";
 import { getServerSession } from "next-auth/next";
 import { authOptions } from "@/app/api/auth/[...nextauth]/route";
-import { isUndefined } from "@/lib";
-import { GPTTokens } from "@/lib/gpt-tokens";
+import { isUndefined, calcTokens } from "@/lib";
 import type { supportModelType } from "@/lib/gpt-tokens";
 import { prisma } from "@/lib/prisma";
 
@@ -78,8 +77,14 @@ const stream = async (
   if (userId) {
     const final = [...messages, { role: "assistant", content: resultContent }];
 
-    const tokenInfo = new GPTTokens({ model, messages: final });
-    const { usedTokens, usedUSD } = tokenInfo;
+    const { usedTokens, usedUSD } = calcTokens(final, model, true);
+
+    console.log("\n\n");
+    console.log(final, "final");
+    console.log(model, "model");
+    console.log(usedTokens, "usedTokens");
+    console.log(usedUSD, "usedUSD");
+    console.log("\n\n");
 
     const findUser = await prisma.user.findUnique({
       where: { id: userId },
@@ -162,13 +167,6 @@ export async function POST(request: Request) {
         messages,
       }),
     });
-
-    if (session) {
-      await prisma.user.update({
-        data: { recentlyUse: new Date() },
-        where: { id: session?.user.id },
-      });
-    }
 
     if (response.status !== 200) {
       return new Response(response.body, { status: 500 });