+++ src/hooks/index.ts
@@ -1,5 +1,4 @@
 export * from "./useChannel";
-export * from "./useChat";
 export * from "./useConfig";
 export * from "./useMobileMenu";
 export * from "./useOpenAI";